Refactor NewsList Component to reuse on the favorite page
