+++ src/components/NewsList/index.jsx
@@ -1,28 +1,9 @@
-import React, { useEffect, useCallback } from "react";
-import useNearScreen from "hooks/useNearScreen";
-import debounce from "just-debounce-it";
-import { useHackNews } from "hooks/useHackNews";
+import React from "react";
 import CardNews from "components/CardNews";
 import Spinner from "components/Spinner";
 import "./NewList.css";
 
-export default function NewsList() {
-  const { newsList, setPage, loading, loadingNextPage } = useHackNews({
-    keyword: "reactjs",
-  });
-  const { isNearFinalElement, finalPageDivRef } = useNearScreen({
-    firstTime: false,
-  });
-
-  const debounceHandleNextPage = useCallback(
-    debounce(() => setPage((previousPage) => previousPage + 1), 500),
-    [setPage]
-  );
-
-  useEffect(() => {
-    if (isNearFinalElement) debounceHandleNextPage();
-  }, [isNearFinalElement, debounceHandleNextPage]);
-
+export default function NewsList({ newsList, loading, loadingNextPage }) {
   return (
     <div className="list-news">
       {loading ? (
@@ -36,7 +17,6 @@ export default function NewsList() {
         ))
       )}
       {loadingNextPage && <Spinner />}
-      <div id="finalPageDivRef" ref={finalPageDivRef}></div>
     </div>
   );
 }
